T5 change int to float

+++ W1/A1_T5.py
@@ -1,8 +1,8 @@
 print("Calculate the area of a wall.")
 Feed = input("Enter the width in metres: ")
-Width = int(Feed)
+Width = float(Feed)
 Feed = input("Enter the height in metres: ")
-Height = int(Feed)
+Height = float(Feed)
 print(f"Width is {Width} m and height is {Height} m.")
 Area = Width * Height
 print(f"The wall will be {Area} square metres.")
